chore: handle conflicts
